Add REST client
In order to query the remote server, a REST client is added. It is
the one that comes with Gerrit's acceptance tests with a modified
constructor in order to avoid more dependencies.

The BUCK build is only supported in Gerrit tree. The support for
standalone BUCK build is already prepared, but will only work with
Gerrit 2.11 where the 'STANDALONE_MODE' flag is available.

Change-Id: Ie12e02a6552643eea122e5fabd71bbd4a3f7ba3e