Keep the import lock file(s) under the @PluginData folder.
When the lock file was kept under the project's Git repository it was
impossible to create a lock before the Git repository existed. This
meant that when two threads tried to import the same non-existing
project, one would succeed with the "git init" and the other one would
fail. However, the failure would be an IOException and not a failure to
acquire a lock.

Keeping the lock file(s) under the @PluginData folder enables us to
create a lock for an import of a non-existing project even before we
perform the "git init". Another advantage is that, by scanning the
@PluginData folder we could easily find non-finished imports and imports
currently running.

Change-Id: I816d6fcd4652c4e1cd594f919d3ce3b91583aa7c