Create changes and replay revisions
This is the first step for replaying changes on the target system. It
only creates the entries in the changes table.

If the owner of a change does not exist in the target system an
exception is thrown and the project import is aborted. This means
that at the moment it is not possible to import a project that has
changes from a user that never logged in into the target system, which
is e.g. a problem if this user has left the company. For now this
problem is ignored, but it must be addressed later in a follow-up
change.

Also for now the user check is kept simple. It may not fit to all
authentication types and may need to be improved later.

Draft changes will only be imported if the user that is used for the
import sees the draft changes, e.g. if it has the 'View Drafts' global
capability.

The change is reindexed so that it can be found by change queries.

At the moment repeating the import is not supported yet. This must be
addressed in a follow-up change.

Change-Id: I4a9d4404015250ccf62214c890e01c9c25b9190b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>