Replay change messages
If the author of a change message does not exist in the target system
an exception is thrown and the project import is aborted. This means
that at the moment it is not possible to import a project that has
change messages from a user that never logged in into the target
system, which is e.g. a problem if this user has left the company. For
now this problem is ignored, but it must be addressed later in a
follow-up change.

Change-Id: I473d75348547e6c1958bcfb7995c311f10a3ea2e
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>