Add approvals to replayed change
If the user of an approval does not exist in the target system an
exception is thrown and the project import is aborted. This means
that at the moment it is not possible to import a project that has
approvals from a user that never logged in into the target system,
which is e.g. a problem if this user has left the company. For now
this problem is ignored, but it must be addressed later in a follow-up
change.

Change-Id: Iaaefd29127e08b6d0c9385c4a6b8696524a5fdd2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>