Set original creation timestamp and uploader on replayed patch sets
Requires [1].

If the uploader of a patch set does not exist in the target system an
exception is thrown and the project import is aborted. This means
that at the moment it is not possible to import a project that has
a patch set that was uploaded by a user that never logged in into the
target system, which is e.g. a problem if this user has left the
company. For now this problem is ignored, but it must be addressed
later in a follow-up change.

[1] https://gerrit-review.googlesource.com/64920

Change-Id: I603048fde3b51e4a22d633d4d618218ed792890a
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>