Replay inline comments
Requires [1,2]

If the author of an inline comment does not exist in the target system
an exception is thrown and the project import is aborted. This means
that at the moment it is not possible to import a project that has
inline comments from a user that never logged in into the target
system, which is e.g. a problem if this user has left the company. For
now this problem is ignored, but it must be addressed later in a
follow-up change.

[1] https://gerrit-review.googlesource.com/65011
[2] https://gerrit-review.googlesource.com/65010

Change-Id: If0b56741350b72c1cdd00bf528afbb39ef7df9db
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>