Automatically create missing accounts if auth type is LDAP
Replaying changes fails if any referenced user is not existing in the
target system. If LDAP is configured as auth type we now try to
authenticate the user and if the authentication is successful a user
account is automatically created, so that the change can be replayed.
It's assumed that source and target system use the same LDAP server.

Change-Id: Ie2cb433fe7e2a8224bfc5b2fb31491c71c6deb8d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>