Fix replaying of revisions when the old and new ref are the same
In the unlikely case when the imported change gets the same number as in
the source system the create/delete of the new/old patch set ref would
fail in the create step as this ref already exists.

Implement a proper ref rename which takes care about this special case.

Change-Id: I889543de11815b171674f170620c43bf8ced54b5