Do not swallow RuntimeExceptions.
Only the declared exceptions mentioned at the catch block were
reported as project import failures. RuntimeExceptions were silently
ignored.

We had also issues with updating the progress monitoring in
this case. The done subtask was triggered a the bottom of the
corresponding try block while the failed subtask was triggered
inside the catch block.

When importing a project any kind of exception is considered
as a failure and reported accordingly since in this case the
imported project might not contain everything it needs.

Change-Id: I415bfc615b09979cbe9b3fc75a856e84cfdedcc7