Fix the NPE when importing merged changes
The "Gerrit Code Review" user, which is owner of some of the change
messages of merged changes doesn't exist as an account. Therefore it
couldn't be resolved.

Handle the "Gerrit Code Review" user as a special case when replaying
change messages. We cannot make use of the ChangeMessageUtil because
instantiating of a ChangeControl for this user is not possible which
would be required for instantiating a ChangeUpdate.

Change-Id: Ie5662259ce358d2ec6ac56f11c575269276eee95
NOTE: this means that the note-db is not updated in this case.