Fix import lock releasing when precondition check fails
The precondition check was done outside of the try..finally which makes
sure to release the import lock. Therefore, after a failed precondition
check the import for this project would remain locked.

Further, when a precondition check fails we write nothing into the
locked file and a subsequent commit was failling because the LockFile.os
was still open (and not null). In this case we should only unlock the
LockFile, without committing.

Also, make sure to only LockFile.commit after something is written into
the file.

Change-Id: I27056a428a703411b60f519aa753e534d74437fa