Let the SSH command and REST import one project in one request
Remove the support for specifying multiple project names in one
command/request. A concurrent import of many projects can still be
achieved by executing multiple import commands from a client
concurrently.

This removes unnecessary complexity from the plugin code and also
provides us with more flexibility for adding new options to the import
command as the options will always be related to one project only.

Since the REST endpoint now only deals with the import of one project,
the project name should be part of the URL, rather than being
contained in the request body. Creation of new REST resources is done
via PUT. This is why the REST endpoint to import a project is now:

  PUT /config/server/importer~projects/<project-id>

This URL scheme is nice since it can easily be extended later, e.g.
to list imported projects:

  GET /config/server/importer~projects/

or to continue a project import:

  PUT /config/server/importer~projects/<project-id>/continue

Having the REST endpoint deal only with one project also simplifies
the error handling. It is not needed anymore to collect error messages
in a message buffer, but errors can be directly mapped to status codes
by throwing REST exceptions.

Also remove the measuring of the import time. If this information is
needed the duration can be retrieved by using the 'time' command on
client side.

Change-Id: I203c73695fda2dec2e3c423ca2eaac719d16ac98