Ensure common validation logic for rest and ssh
When rest was used for a project import the validation logic of the
importer plugin raised an exception when the parent project provided
from the call did not exist.

When ssh was used for a project import the validation logic of the
importer plugin was not used since we got an error earlier before the
parent got injected into the plugin.

Now the validation logic of the plugin is called for both, ssh and
rest in the same way.

Change-Id: I1d08c260f00a5c65854e9518855af2636bff5d35