Allow to import project with other name
This way a project can be renamed while importing.

The name that is locked is the target name. This allows to import the
same source project multiple times, but with different names.
Due to this the project name in the REST endpoints is always the
target name. The source name can be provided in the JSON body. If it
is not specified it is assumed to be the same as the target name.

Change-Id: I1eef75cc2d77885cc065f384ec22205497f2a3b5
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>