Add REST endpoint to copy a project
There is a new REST endpoint to copy a project within the same Gerrit
server:

  PUT /projects/<project-to-be-copied>/copy

The name of the target project must be specified in the JSON in the
request body.

When copying a project the parent project cannot be changed. This is
intentional since changing a parent project is an administrator task
whereas the permission to copy projects may be assigned to
non-administrators.

There is a new capability to allow copying projects. The import
capability was not reused because importing a project from another
system is more "evil" than copying a project within one Gerrit server.
Hence the capability to copy projects may be assigned to more people
than the import capability.

Copying is already possible by using the import REST endpoint,
however the new copy REST endpoint is more convinient since one
doesn't need to provide URL, user and password of the source system.
Internally the copy REST endpoint just delegates to the import REST
endpoint.

Change-Id: Id103356ffa989d4524ee61c10fe269ed341f0df8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>