Fetch ssh keys for users involved in the project migration.
Status before this change:
When a user does not already exist in the target system  it is created
without transfering the ssh keys from the source system into the
target system.

Status after this change:
The user is created and the ssh keys are transferred.

The ssh keys are only transferred when the user does not already exist
in the target system. For already existing users in the target system
the ssh keys are not transferred.

Change-Id: I00654105583c556b7440f29d2cc086d7e95a01de