Add backend functionality needed to import groups
With this commit, imported groups are actually persisted in the target
system.

Import of groups in a recursive way is not supported, i.e. owner and
member groups need to be already existent on the target system,
otherwise the import fails. A possible self-ownership relation of a
group is not considered to be a recursive import and is supported.

Change-Id: I3a291785f6201cf03a61a84940bcea370f83032c