On resume skip unmodified changes
When a change is imported it gets the same last modified timestamp as
the change in the source system. Any modification of the change in the
source system would update the last modified timestamp. For resuming
an import this means that we can skip all changes were the last
modified timestamp in the target system is the same as in the source
system. Skipping the unmodified changes means that we only get the
"Resumed import of ..." change message if the change was modified
since the last import.

Change-Id: I20f9b1d2bad42d9479e36422a777b6e628527bcd
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>