Do not fail resume if there are new patch sets in target system
Resuming an import after doing changes in the target system is not
supported, still we can be nice and handle a few cases.

At the moment resuming the import fails if a change in the target
system has a new patch set. It fails because we try to fetch inline
comments for this patch set from the source system and this request
fails with '404 Not Found'. Recognize this case and continue the
import.

Change-Id: I547c9f6061b487be3dfd63eb4df6fcaca009a54d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>