Add option to force resume project import
Normally changes that have the same last modified timestamp in the
source and target system are skipped on resume. This is good since it
makes the resume faster and it avoids unneeded "Resumed import of ..."
messages on the changes. However in some cases these changes should
not be skipped. E.g. approvals for an unknown label are ignored during
an import, if later this label is configured on the target server, the
skipped approvals will still not appear in the target system after
resume if the change is skipped due to not having been modified. In
this case the resume can now be done forcefully.

Change-Id: I6be084dbed00d5e530beaa3904eb687a5c5fb849
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>