[Fix] list project fails due to attempt to read folders
When the list of all already imported projects is created we traverse
the import files located in the data/importer directory.
Up to now also folders was taken into account. Since folders can't be
read as json the construction of the list failed.
With this change only files not ending with .lock in the file name are
taken into account.

Change-Id: I14e9c797d9963040ea0c08da2c35ae535c5c44ea