Use local API for doing project copy
So far when doing a project copy "remote" requests from the Gerrit
server to itself were done. Due to this the importing user was
required to have a generated HTTP password.

There is now a local API that is used to do a project copy so that in
this case there are no "remote" calls anymore.

Change-Id: I2112697753c92350ba4cad7a5eb4ca51aa2b8401
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>