[fix] calculate project name out of import status file
The file name of the import project file was used as project name.
This does not take into account that project names might be organized
in an hierarchical way.

The project name is now the delta between the data/importer folder
and the import file.

Up to now the file name was used as project name. But projects might be
organized in an hierarchical way. Hence it is not sufficient to use the
file name of the import status file.

Now the path between the data folder of the importer and the import
status file is used as project name.

This fixes e.g. incomplete project names in the import project view.

Change-Id: If76d87d8f30c35929f03315959e6bd203a4a848b