Fail with proper error if account has no username or no email
Accounts can only be looked up by username, but some accounts may not
have a username. Recognize this case and fail with a proper error.

If an account exists in the target system it is verified by comparing
the email. This fails if an account has no email, but it now results
in a proper error.

Also mention in the documentation that it is a precondition for the
import that accounts have a username and an email.

Change-Id: Id63e6c9372677ec4a021cf656a0de3937d3b671a
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>