Automatically create missing owner groups on group import
So far the group import was failing if the owner group of the imported
group did not exist in the target system. Now there is an option to
automatically create missing owner groups.

When creating a group on import, it is first created as self-owned
group. Then the owner group is created if it is missing and then the
owner group is assigned to the imported group. This is done to prevent
an infinite loop when there is a circle in the group ownerships.

When importing groups on project import, missing owner groups are
always imported automatically.

Change-Id: I626fcde4a8f158f19fea59267a3c917348f221c8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>