Delete import refs on completion of a project import
After an import is completed the refs in the 'refs/imports/' namespace
are not needed anymore. They rather disturb when an imported project
is copied or imported into another server, because then we have the
old import refs under 'refs/imports/imports/' in the new project.

Explain in the documentation that refs are fetched into the
'refs/imports/' namespace and that the complete import command deletes
them.

Change-Id: I7ebce7dc7631de9e2356605d0f6e97fda1a4c367
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>