Check version of Gerrit source server on project import
Importing a project requires that the source Gerrit server is at
version 2.11 or newer. Running the import against older servers
will result in various errors in the middle of the import. This is why
it's better to check the version of the source Gerrit server early and
fail with a proper error message if it's too old.

Change-Id: I67fb880d969c374fc6ec2948515dc6e73f82afb7
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>