Match the whole project name when listing projects
Project names can contain many path components like:

  relative/path/to/my/project

When traversing the files under the lockRoot, match the whole relative
path of the file against the given match. For example:

  /path/to/lock/root/relative/path/to/my/project

For this file we match the "relative/path/to/my/project" against the
given match and not just the file name "project".

Change-Id: I80e2feff071b2b1a42811ba882aa8cd7b301da8f