Support password piping also for resume import
Up to now password piping was only supported for import project, not
for resume import. In case of a failure during resume import the
password was contained in plain text in the error log. When password
piping is not used this is still the case, but now it is possible to
protect the password by using password piping.

According to the plugin documentation the password piping for resume
project import should already been implemented. But in fact it was
not.

Change-Id: I1b8f5d209a6acbb11eef202d6f490fda1cd180ac