Handle changes with non-existing current revision
We have a few changes that have a current revision that does not match
the revision of any of the patch sets in the change. This is a data
inconsistency and as result the import of such a change fails. Be more
tolerant, log a warning and use the last revision as current patch
set in this case so that this import is not failing.

Change-Id: Ifda86d116d1147362d7b3312d0e11e7a2bd54cc1
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>