Do not skip changes without current revision
If the current revision of a change is a draft patch set and the
querying user cannot see the draft patch set, then the
'current_revision' field in the ChangeInfo is not set. At the moment
we skip such changes, which is bad. Instead use the last revision as
current patch set in this case and log a warning.

Change-Id: I480d70e99079d4edd5ab1b101392d8bc65c3b726
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>