Remove Maven build
The Maven build is packaging too much stuff into the plugin jar.
Building the plugin with Maven results in a plugin jar > 18MB while
with Buck it is < 2MB. Instead of fixing the Maven build, just remove
it. It's anyway too much work to support 2 build systems in the long
run.

Change-Id: I8dffdd7281a3e71e36c54dd26c33d68187d1e043
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>