[fix] only 500 changes are imported.
Via rest api the number of changes returned for a rest query for
changes is limited. The default limit is 500 changes.
Up to now we rely on that default, hence not all changes are imported.

Now we fetch all changes. The number of changes returned by the rest
call is counted. In case more changes are indicated we fetch another
bundle providing the appropriate offset of changes already imported.
This continues until all changes are imported.

Change-Id: I6a8fb85832e20d3f8e26f757b4163f6440b7ad1d