Include '/a' into HTTP repository URLs to trigger authentication
Not having the '/a' prefix may lead to problems when a part of the
project is visible to Anonymous Users. In this case authentication is
never triggered. This is because for HTTP the first request is always
anonymous and then the server must say that authentication is
required. Only then the client can send user credrentials. If a
project is (partly) visible to Anonymous Users then the first
anonymous request is always successful and authentication never
happens. If the user is not authenticated refs which are not visible
to Anonymous Users, but for which the user has read access, are not
visible. As result the importer plugin cannot fetch these refs and
fails during the import with a NullPointerException when the ref for a
change cannot be found. Often the 'refs/meta/config' branch is
affected since this branch is normally not visible for Anonymous
Users.

Change-Id: I483cffded0dfe3077d4cf95d6002900a9837ff9a
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>