Create local user if user does not exist anymore in LDAP
So far, importer plugin aborted  the importing process with an exception
if an user listed as a change owner, reviewer or group member does not
exist anymore in LDAP. That use case arise, for example, when an user
leave the company. The same problem occurs when dealing with non
existing Gerrit internal users.

Create a local user account to replace missing LDAP or service users;
these users are created adding an "_imported" suffix to their username
to be able to recognize them easily.

Change-Id: Iad1b1c3963cca7dff6ebcc59efcf2b5620e0fd42