Create local user if user does not exist anymore in LDAP
So far, importer plugin aborted the importing process with an exception
if an user listed as a change owner, reviewer or group member did not
exist anymore in LDAP. That use case arises, for example, when an user
leaves the company. The same problem occurred when dealing with non
existing Gerrit internal users.

Create local inactive user accounts to replace missing LDAP users and
add them to an special "Imported Users" group. Inactive users are still
able to replay changes. Create Gerrit internal users if they have not
been created previous to the import process.

Change-Id: Iad1b1c3963cca7dff6ebcc59efcf2b5620e0fd42