Update patch set 1
Patch Set 1:

> It would be really nice if we had CI reporting these breakages at
 > review time :(

Totally agreed!

But to detect this kind of breakage during review, we'd need to build
all plugins for each change in gerrit core.
Such a build with clean caches takes ~1 hour on my server [1].
That's ok for nightlies, but really slow when trying to run more often.

But given that this kind of breakage does not happen too often, and we'll
currently notice breakage at least within a day, I'd argue that while the
current situation is not ideal, it's ok.

---------------------------

[1] I clean buck caches before the build, as I've been bitten by
inconsistencies caused by buck's caches getting in the way. That's a bit
hand-wavy, because I cannot reliably reproduce those caching issues. But
before cleaning the cache, build $SOMETIMES failed with error messages
that indicated that *.java and corresponding *.class files do not match.
Nuking the cache always fixed the issue for me.

Patch-set: 1