Update patch set 2
Patch Set 2:

It's just we are hard coding the http library version in the plugin. Replacing this with the one provided by Gerrit it fixes the issue. May be the change I did it's a bit overkill, but for sure, the BUCK files for this plugin need some updates.

Patch-set: 2
Label: Verified=0