Update patch set 1
Patch Set 1:

> > If we do not add those dependencies, the importer plugin won't
 > > compile with the latest version of bucklets.
 > > Do we have other way to work around this?
 > 
 > User the right bucklets version and not the latest one?

I know that using version prior to "fa16504 gerrit_plugin: Pull in the dependencies for GWT 2.8-rc2" will work and not require this change but I am wondering how is the user supposed to know what version to use?

If we do not want to not update stable branches with such changes, maybe we should use bucklets as a git submodule and pin the version.

Patch-set: 1