Update patch set 1
Patch Set 1:

Should we merge that one first and rebase this one on top of it? The only issue with this approach is that even if that change is merged, the plugin won't compile against master given there are other things that need to be adapted (covered in this change), so we'll have (until this change is merged) a non compiling plugin. But that could be OK given that's the case right now.

Patch-set: 1
Reviewer: Hector Oswaldo Caballero <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>