Handle changes in refs/meta
When there is a change for review in branch other than refs/heads/ or
refs/tags, they are presented with the refs/ prefix. This will cause
branch location recorded incorrectly during import.

Detect refs/ prefix instead of refs/heads prefix to fix this.
refs/tags should be okay as it is processed somewhere else.

Change-Id: I7a9b9a4f710398e52a89e21c64e76bcebaf3e1c6
Signed-off-by: Ardo Septama <aseptama@gmail.com>