Update patch set 3
Patch Set 3:

I've built and installed this in a test server running head of stable-2.15.  The plugin loads and can be used from the UI, but it's not importing projects properly.  The project gets created but changes do not get imported.

Shall we merge this as-is and leave that to be fixed later?  I have not tried it on any other version so don't know if it was broken before, it's a regression in this change, or I'm simply doing something wrong.

Patch-set: 3
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>